requirements.txt
langchain
PyGithub
openai
chromadb
python-dotenv
tiktoken
requests
beautifulsoup4
langchain-text-splitters
numpy